added ErrorToast Component
